+++ src/scenario/index.ts
@@ -4,4 +4,4 @@ export * from './route-timetable'
 export * from './buffer-passes'
 export * from './stop-clusters'
 export * from './phases'
-export * from './reportTables'
+export * from './report-tables'
